fix: cast LogLevel to int for ClickHouse serialization and bump version to 10.0.12

+++ src/Aiursoft.ClickhouseLoggerProvider/LoggingDbContext.cs
@@ -24,7 +24,7 @@ public class LoggingDbContext : ClickhouseDbContext
         Logs = new ClickhouseSet<LogEntry>(GetConnection, options.CurrentValue.TableName, log => new object[] 
         {
             log.EventTime,
-            log.LogLevel,
+            (int)log.LogLevel,
             log.Category,
             log.Message,
             log.Exception
